% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno1kg}
\alias{anno1kg}
\title{a data.frame with ancestry information on 1000 genomes samples}
\format{
An object of class \code{data.frame} with 3500 rows and 6 columns.
}
\usage{
data(anno1kg)
}
\description{
a data.frame with ancestry information on 1000 genomes samples
}
\keyword{datasets}
