% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filt.R
\name{biplot_filtered}
\alias{biplot_filtered}
\title{create a filtered biplot}
\usage{
biplot_filtered(pca, pc = c(1, 2), top_n = NULL, min_norm = NULL)
}
\arguments{
\item{pca}{prcomp output}

\item{pc}{2-vector, which pcs to use}

\item{top_n}{numeric, number of features to use for directional arrows over projection}

\item{min_norm}{lower bound on L2 norm of feature loadings to use}
}
\description{
create a filtered biplot
}
