% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch_19to38}
\alias{ch_19to38}
\title{a liftOver chain as defined by rtracklayer, for hg19 to hg38}
\format{
An object of class \code{Chain} of length 25.
}
\usage{
data(ch_19to38)
}
\description{
a liftOver chain as defined by rtracklayer, for hg19 to hg38
}
\keyword{datasets}
